+++ crates/biome_js_formatter/src/js/lists/object_member_list.rs
@@ -1,4 +1,4 @@
-use crate::context::trailing_comma::FormatTrailingComma;
+use crate::context::trailing_commas::FormatTrailingCommas;
 use crate::prelude::*;
 use biome_js_syntax::JsObjectMemberList;
 use biome_rowan::{AstNode, AstSeparatedList};
@@ -10,7 +10,7 @@ impl FormatRule<JsObjectMemberList> for FormatJsObjectMemberList {
     type Context = JsFormatContext;
 
     fn fmt(&self, node: &JsObjectMemberList, f: &mut JsFormatter) -> FormatResult<()> {
-        let trailing_separator = FormatTrailingComma::ES5.trailing_separator(f.options());
+        let trailing_separator = FormatTrailingCommas::ES5.trailing_separator(f.options());
 
         let mut join = f.join_nodes_with_soft_line();