+++ crates/biome_js_formatter/src/ts/lists/enum_member_list.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
 
-use crate::context::trailing_comma::FormatTrailingComma;
+use crate::context::trailing_commas::FormatTrailingCommas;
 use biome_js_syntax::TsEnumMemberList;
 
 #[derive(Debug, Clone, Default)]
@@ -10,7 +10,7 @@ impl FormatRule<TsEnumMemberList> for FormatTsEnumMemberList {
     type Context = JsFormatContext;
 
     fn fmt(&self, node: &TsEnumMemberList, f: &mut JsFormatter) -> FormatResult<()> {
-        let trailing_separator = FormatTrailingComma::ES5.trailing_separator(f.options());
+        let trailing_separator = FormatTrailingCommas::ES5.trailing_separator(f.options());
         let mut joiner = f.join_nodes_with_soft_line();
 
         for variant in node